+++ microservices-project/payment-service/src/main/java/com/lisi/booknavigator/paymentservice/event/PaymentEvent.java
@@ -10,7 +10,7 @@ import lombok.*;
 @Getter
 @Setter
 public class PaymentEvent {
-    private Integer  paymentId;
+    private Long  paymentId;
     private EventType eventType;
     private Payment payment;
